Expose file count/size on content mgmt page
