+++ kolibri/plugins/management/assets/src/views/manage-content-page/index.vue
@@ -40,20 +40,26 @@
         <thead>
           <tr>
             <th class="col-header col-channel" scope="col"> Channel Name </th>
-            <!-- <th class="col-header col-export" scope="col"> # of Contents </th>
+            <th class="col-header col-export" scope="col"> # of Contents </th>
             <th class="col-header col-export" scope="col"> Size </th>
-            <th class="col-header col-export" scope="col"> Last updated </th> -->
+            <!-- <th class="col-header col-export" scope="col"> Last updated </th> -->
           </tr>
         </thead>
         <!-- Table body -->
         <tbody>
           <tr v-for="channel in channelList">
             <!-- Channel Name -->
-            <th scope="row" class="table-cell" width="70%">
+            <td scope="row">
               <span class="channel-name">
                 {{ channel.title }}
               </span>
-            </th>
+            </td>
+            <td>
+              {{ numberOfFilesInChannel(channel.id) }}
+            </td>
+            <td>
+              {{ totalSizeOfFilesInChannel(channel.id) }}
+            </td>
             <!-- Export Button -->
 <!--             <td class="table-cell table-export" width="30%">
             </td> -->
@@ -71,6 +77,7 @@
 
   const actions = require('../../state/actions');
   const ContentWizardPages = require('../../constants').ContentWizardPages;
+  const bytesForHumans = require('./bytesForHumans');
 
   module.exports = {
     $trNameSpace: 'manageContentState',
@@ -96,6 +103,20 @@
     },
     destroyed() {
       clearInterval(this.intervalId);
+    },
+    methods: {
+      numberOfFilesInChannel(channelId) {
+        if (this.pageState.channelInfo[channelId]) {
+          return this.pageState.channelInfo[channelId].numberOfFiles;
+        }
+        return 'Please wait...';
+      },
+      totalSizeOfFilesInChannel(channelId) {
+        if (this.pageState.channelInfo[channelId]) {
+          return bytesForHumans(this.pageState.channelInfo[channelId].totalFileSizeInBytes);
+        }
+        return 'Please wait...';
+      }
     },
     computed: {
       wizardComponent() {
@@ -189,12 +210,6 @@
     color: $core-text-annotation
     font-weight: normal
     font-size: 80%
-
-  .col-channel
-    width: 90%
-
-  .col-export
-    width: 10%
 
   .table-cell
     font-weight: normal // compensates for <th> cells
